Moved ErrorMatrixTag to DataFormats/Common
