+++ DataFormats/GeometrySurface/interface/LocalErrorExtended.h
@@ -2,7 +2,7 @@
 #define _TRACKER_LOCAL_ERROR_H_
 
 #include "DataFormats/GeometrySurface/interface/LocalErrorBaseExtended.h"
-#include "DataFormats/GeometryCommonDetAlgo/interface/ErrorMatrixTag.h"
+#include "DataFormats/Common/interface/ErrorMatrixTag.h"
 
   /**
    * typedef to a  GlobalErrorBase object defined as a 3*3 covariance matrix
